FileUploadTraitを追加し、cfFormMailerクラスにファイルアップロード機能を統合
