+++ includes/class.cfFormMailer.inc.php
@@ -17,6 +17,7 @@ require_once __DIR__ . '/traits/TemplateMethodsTrait.php';
 require_once __DIR__ . '/traits/MailMethodsTrait.php';
 require_once __DIR__ . '/traits/FormRenderingTrait.php';
 require_once __DIR__ . '/traits/UtilityMethodsTrait.php';
+require_once __DIR__ . '/traits/FileUploadTrait.php';
 
 class Class_cfFormMailer
 {
@@ -26,6 +27,7 @@ class Class_cfFormMailer
     use MailMethodsTrait;
     use FormRenderingTrait;
     use UtilityMethodsTrait;
+    use FileUploadTrait;
 
     public $cfg = array();
 
@@ -327,27 +329,6 @@ class Class_cfFormMailer
         return $text;
     }
 
-    private function extension($tmp_name)
-    {
-        $info = @getimagesize($tmp_name);
-        if ($info === false || !isset($info['mime'])) {
-            return '';
-        }
-        return '.' . $this->_getType($info['mime']);
-    }
-
-    private function cfm_upload_tmp_name($tmp_name)
-    {
-        $tmp_path = CFM_PATH . 'tmp';
-        if (!is_dir($tmp_path) && !mkdir($tmp_path)) {
-            exit('ディレクトリを生成できません');
-        }
-        if (!is_writable($tmp_path)) {
-            exit('ディレクトリに書き込めません');
-        }
-        return $tmp_path . substr($tmp_name, strlen(dirname($tmp_name)));
-    }
-
     private function getToken()
     {
         return base_convert(str_shuffle(mt_rand()), 10, 36);
@@ -486,21 +467,6 @@ class Class_cfFormMailer
 
         $rs = $this->sendAutoReply();
         return $rs;
-    }
-
-    public function cleanUploadedFiles()
-    {
-        $uploaded = sessionv('_cf_uploaded');
-        if (empty($uploaded)) {
-            return;
-        }
-        if (!is_array($uploaded)) {
-            return;
-        }
-        foreach ($uploaded as $attach_file) {
-            unlink($attach_file['path']);
-        }
-        unset($_SESSION['_cf_uploaded']);
     }
 
     /**
@@ -900,63 +866,6 @@ class Class_cfFormMailer
             "/^(?:[a-z0-9+_-]+?\.)*?[a-z0-9_+-]+?@(?:[a-z0-9_-]+?\.)*?[a-z0-9_-]+?\.[a-z0-9]{2,5}$/i",
             $addr
         );
-    }
-
-    /**
-     * ファイルのMIMEタイプを取得
-     * class.upload.php 使用推奨
-     *
-     * @param string  $filename MIMEタイプを調べるファイル
-     * @param string  $field  アップロードされたフィールド名
-     * @return string MIMEタイプ。失敗の場合は false
-     */
-    private function _getMimeType($filename, $field = '')
-    {
-        if (class_exists('upload')) {
-            // class.upload.php使用
-            $up = new upload($filename);
-            if ($up->uploaded) {
-                return $up->file_src_mime;
-            }
-            return false;
-        }
-
-        // class.upload.php未使用。image以外の結果はあまり信用できない
-        $size = @getimagesize($filename);
-        if ($size === false) {
-            if (isset($_FILES[$field]['type']) && $_FILES[$field]['type']) {
-                return $_FILES[$field]['type'];
-            }
-            return false;
-        }
-
-        return $size['mime'];
-    }
-
-    /**
-     * MIMEタイプからファイルの識別子を取得
-     *
-     * @param string  MIMEタイプ
-     * @return  string  タイプ
-     */
-    private function _getType($mime)
-    {
-        $mime_list = array(
-            'image/gif'          => 'gif',
-            'image/jpeg'         => 'jpg',
-            'image/pjpeg'        => 'jpg',
-            'image/png'          => 'png',
-            'application/pdf'    => 'pdf',
-            'text/plain'         => 'txt',
-            'text/html'          => 'html',
-            'application/msword' => 'word'
-        );
-
-        if (isset($mime_list[$mime])) {
-            return $mime_list[$mime];
-        }
-
-        return '';
     }
 
     private function charset()
